Implemented restructuring
