+++ src/components/Card.css
@@ -3,8 +3,9 @@
     border-radius: 8px;
     box-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.14),
       0 1px 5px 0 rgba(0, 0, 0, 0.12), 0 3px 1px -2px rgba(0, 0, 0, 0.2);
-    padding: 24px;
-    width: 720px;
+    padding: 1em;
+    height: 150px;
+    width: 450px;
 }
 
 @media screen and (max-width: 600px) {
@@ -12,6 +13,5 @@
       height: 70%;
       padding: 15px;
       display: grid;
-      grid-template-rows: auto;
   }
 }
