+++ src/components/Container.js
@@ -1,9 +1,26 @@
 import React from "react";
+import Logo from "./Logo";
+import LocationInfo from "./location/LocationInfo";
+import Timer from "./Timer";
+import CurrentWeather from "./weather/CurrentWeather";
+import HourlyWeather from "./weather/HourlyWeather";
+import ForecastContainer from "./weather/ForecastContainer";
+import Radar from "./weather/Radar";
 
 import "./Container.css";
 
-const Container = () => {
-  return <div className="container" />;
+const Container = ({ city, location, weatherData }) => {
+  return (
+    <div className="container">
+      <Logo />
+      <LocationInfo city={city} location={location} />
+      <Timer />
+      <CurrentWeather weatherData={weatherData} />
+      <HourlyWeather />
+      <ForecastContainer />
+      <Radar />
+    </div>
+  );
 };
 
 export default Container;
