+++ src/components/Logo.css
@@ -1,5 +1,5 @@
 .logo {
-    font-size: 48rem;
+    font-size: 56px;
     font-family: Verdana, Geneva, Tahoma, sans-serif;
-    color: white;
+    color: black;
 }
