+++ src/components/location/LocationInfo.css
@@ -1,43 +1,29 @@
 .location-info {
+    display: flex;
     align-items: center;
 
 }
 .city {
     font-size: 24px;
     font-weight: 500;
-    margin-bottom: 16px;
-    margin-left: 10px;
 }
 
 .location {
     font-size: 24px;
     font-weight: 500;
-    margin-left: 10px;
 }
 
 @media screen and (max-width: 600px) {
     .location-info {
-        display: grid;
-        grid-template-rows: repeat(2);
-        justify-content: center;
-        align-items: center;
+        font-size: 20px;
 
     }
     .city {
         font-size: 20px;
-        grid-row: 2;
-        margin-bottom: 0px;
-        margin-left: 15px;
-        margin-top: -30px;
-        justify-content: center;
     }
     
     .location {
-        font-size: 60px;
-        grid-row: 1;
-        margin-bottom: 0px;
-        margin-left: 0px;
-        align-self: center;
+        font-size: 20px;
     }
