+++ src/components/location/LocationInfo.js
@@ -8,6 +8,7 @@ const LocationInfo = ({ city, location }) => {
   return (
     <div className="location-info">
       <div className="city">{city}</div>
+      <div className="slash">/</div>
       <div className="location">{location}</div>
     </div>
   );
