+++ src/components/weather/DailyForecast.css
@@ -1,0 +1,3 @@
+.daily-wrapper{
+    flex-basis: 200px;
+}
