+++ src/components/weather/ForecastContainer.css
@@ -1,0 +1,7 @@
+.forecast-wrapper {
+    display: flex;
+    flex: auto;
+    flex-direction: row;
+    gap: 1em;
+    padding: 1em; 
+}
