+++ src/components/weather/HourlyWeather.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Card from "../Card";
 
-import "HourlyWeather.css";
+import "./HourlyWeather.css";
 
 // Need to upgrade API and figure out what param to bring in via props
 // Then need to figure out structure, thinking date on top, icon, middle, highs and lows bottom
